Add negative assignment test cases to parser
Also fix a bug where if we failed to have a newline or EOF after an
assignment then we would double free some memory. It was because we had
already inserted the key and value into the map but also errdefered
freeing the key and value. That meant when we freed the table we would
double free them.