Heap allocate the top-level table
This is so that we can have `Parser.current_table` be a pointer to the
current table. We need a stable address for any table which means we have
to heap allocate the top-level one or make sure `Parser` isn't moved.
I've opted for the former here as it is harder to misuse.

Currently `Parser.current_table` and `Parser.top_level_table` are always
the same, but that will change with nested tables and dotted field
access.