Parse dotted fields
To do so we need to handle creating tables within tables. `parseKey` now
allows a '.' after a key. If this happens it sets `current_table` to a
table for the given key, which is created if necessary. It then calls
itself recursively to do the assignment. Once that has been done it puts
the previous `current_table` back.