Lex and parse integers
As anything numeric is now considered a number (rather than a key) we
lose the ability to have numeric keys without quoting them. This could
be remedied by having the lexer know when it has just had a newline
(in which case the "number" must actually be a key) or have the parser
accept a number as the first thing on a line and turn it back into a
string to use as the key.