Ensure lexer gets destructed
Before this commit `Lexer`'s arena would not be freed unless we got to
the end of lexing (i.e. we called `Lexer.next` until it was null). This
would not happen if we failed parsing for example. Sometimes it works
out to be fine - i.e. if the iteration we failed on had not allocated
any memory.

To fix simply add a deinit method on `Lexer` and call it from the
parser.